Reorganization and cleanup
