+++ dfno/__init__.py
@@ -1,2 +1,3 @@
-from .utils import *
 from .dfno import *
+from .loss import *
+from .utils import *
