+++ training/navier_stokes/requirements_navier_stokes.txt
@@ -1,0 +1,3 @@
+mat73
+matplotlib
+scipy
