+++ training/two_phase/train_two_phase.py
@@ -1,37 +1,22 @@
-import distdl, torch, os, time, h5py
+import torch, os, time, h5py
 import numpy as np
 import azure.storage.blob
 import matplotlib.pyplot as plt
-from mpi4py import MPI
-from pfno import ParallelFNO, create_standard_partitions
-from distdl.backends.mpi.partition import MPIPartition
-from sleipner_dataset import DistributedSleipnerDataset3D
-from distdl.nn.repartition import Repartition
-from loss import DistributedRelativeLpLoss
 
+from dfno import DistributedFNO, create_standard_partitions, DistributedRelativeLpLoss, get_env
 from dotenv import load_dotenv
-from loss import DistributedRelativeLpLoss
+from sleipner_dataset import DistributedSleipnerDataset3D
+
+# Load dataset information from environment
+load_dotenv()
 
 # Partitions
 n = 4
 P_world, P_x, P_root = create_standard_partitions((1, 1, 1, n, 1, 1))
-num_gpus = n
-device_ordinal = P_x.rank % num_gpus
+use_cuda, cuda_aware, device_ordinal, device, ctx = get_env(P_x, num_gpus=n)
 dtype = torch.float32
 
-# this fails when placed at the top?? Torch must be setting cuda version or something...
-try:
-    device = torch.device(f'cuda:{device_ordinal}')
-    import cupy
-    ctx = cupy.cuda.Device(device_ordinal)
-except:
-    device = torch.device('cpu')
-    from contextlib import nullcontext
-    ctx = nullcontext()
-
 with ctx:
-    # Load env vars
-    load_dotenv()
 
     # Reproducibility
     torch.manual_seed(P_x.rank + 123)
@@ -39,15 +24,15 @@ with ctx:
 
     # Data dimensions
     nb = 1
-    shape = (60, 60, 64, 10)    # X Y Z T
-    num_train = 1
-    num_valid = 1
+    shape = (60, 60, 64, 30)    # X Y Z T
+    num_train = 800
+    num_valid = 200
 
     # Network dimensions
     channel_in = 2
     width = 20
     channel_out = 1
-    modes = (12, 12, 12, 4)
+    modes = (12, 12, 12, 8)
 
     # Data store
     container = os.environ['CONTAINER']
@@ -65,8 +50,7 @@ with ctx:
             normalize=True)
 
     # Validation dataset
-    #val_idx = torch.linspace(0, num_train+num_valid, num_valid, dtype=torch.int32).long()
-    val_idx = train_idx
+    val_idx = torch.linspace(num_train+1, num_train+num_valid, num_valid, dtype=torch.int32).long()
     valid_data = DistributedSleipnerDataset3D(P_x, val_idx, client, container, data_path, shape,
             normalize=True)
 
@@ -76,7 +60,7 @@ with ctx:
     P_world._comm.Barrier()
 
     # FNO
-    pfno = ParallelFNO(
+    dfno = DistributedFNO(
             P_x,
             [nb, channel_in, *shape[:-1], 1],
             shape[-1],
@@ -89,10 +73,9 @@ with ctx:
     # Training
     num_epochs = 100
     checkpoint_interval = 10
-    out_dir = '/home/azureuser/dfno/pfno/data'
-    #out_dir = '/datadrive/thomas'
+    out_dir = 'data/'
 
-    parameters = [p for p in pfno.parameters()]
+    parameters = [p for p in dfno.parameters()]
     #criterion = distdl.nn.DistributedMSELoss(P_x).to(device)
     criterion = DistributedRelativeLpLoss(P_x).to(device)
     if len(parameters) > 0:
@@ -109,7 +92,7 @@ with ctx:
     for i in range(num_epochs):
 
         # Loop over training data
-        pfno.train()
+        dfno.train()
         train_loss = 0
         n_train_batch = 0
 
@@ -121,7 +104,7 @@ with ctx:
             x = x.to(device)
             y = y.to(device)
 
-            y_hat = pfno(x)
+            y_hat = dfno(x)
             loss = criterion(y_hat, y)
             if P_root.active:
                 #print(f'epoch = {i}, batch = {j}, loss = {loss.item()}')
@@ -144,7 +127,7 @@ with ctx:
         P_x._comm.Barrier()
 
         # Loop over validation data
-        pfno.eval()
+        dfno.eval()
         valid_loss = 0
         n_valid_batch = 0
 
@@ -154,7 +137,7 @@ with ctx:
                 x = x.to(device)
                 y = y.to(device)
 
-                y_hat = pfno(x)
+                y_hat = dfno(x)
                 loss = criterion(y_hat, y)
                 if P_root.active:
                     #print(f'epoch = {i}, batch = {j}, test loss = {loss.item()}')
@@ -175,15 +158,15 @@ with ctx:
                 fid.create_dataset('train_loss', data=train_accs)
                 fid.create_dataset('valid_loss', data=valid_accs)
                 fid.close()
-                #print(f'rank = {P_x.rank}, saved loss: {lossname}')
+                print(f'rank = {P_x.rank}, saved loss: {lossname}')
 
             model_path = os.path.join(out_dir, f'model_{(i+1):04d}_{P_x.rank:04d}.pt')
-            torch.save(pfno.state_dict(), model_path)
-            #print(f'rank = {P_x.rank}, saved model: {model_path}')
+            torch.save(dfno.state_dict(), model_path)
+            print(f'rank = {P_x.rank}, saved model: {model_path}')
 
     # Save after training
     model_path = os.path.join(out_dir, f'model_{P_x.rank:04d}.pt')
-    torch.save(pfno.state_dict(), model_path)
+    torch.save(dfno.state_dict(), model_path)
     print(f'rank = {P_x.rank}, saved model after final iteration: {model_path}')
 
     if P_root.active:
